+++ CopilotKit/packages/react-core/src/components/copilot-provider/copilotkit.tsx
@@ -45,6 +45,10 @@ import { useCopilotRuntimeClient } from "../../hooks/use-copilot-runtime-client"
 import { shouldShowDevConsole } from "../../utils";
 import { CopilotErrorBoundary } from "../error-boundary/error-boundary";
 import { Agent, ExtensionsInput } from "@copilotkit/runtime-client-gql";
+import {
+  LangGraphInterruptAction,
+  LangGraphInterruptActionSetterArgs,
+} from "../../types/interrupt-action";
 
 export function CopilotKit({ children, ...props }: CopilotKitProps) {
   const showDevConsole = props.showDevConsole === undefined ? "auto" : props.showDevConsole;
@@ -342,6 +346,24 @@ export function CopilotKitInternal(cpkProps: CopilotKitProps) {
 
   const showDevConsole = props.showDevConsole === undefined ? "auto" : props.showDevConsole;
 
+  const [langGraphInterruptAction, _setLangGraphInterruptAction] =
+    useState<LangGraphInterruptAction | null>(null);
+  const setLangGraphInterruptAction = useCallback((action: LangGraphInterruptActionSetterArgs) => {
+    _setLangGraphInterruptAction((prev) => {
+      if (prev == null) return action as LangGraphInterruptAction;
+      if (action == null) return null;
+      let event = prev.event;
+      if (action.event) {
+        // @ts-ignore
+        event = { ...prev.event, ...action.event };
+      }
+      return { ...prev, ...action, event };
+    });
+  }, []);
+  const removeLangGraphInterruptAction = useCallback((): void => {
+    setLangGraphInterruptAction(null);
+  }, []);
+
   return (
     <CopilotContext.Provider
       value={{
@@ -388,6 +410,9 @@ export function CopilotKitInternal(cpkProps: CopilotKitProps) {
         setAuthStates_c: setAuthStates,
         extensions,
         setExtensions,
+        langGraphInterruptAction,
+        setLangGraphInterruptAction,
+        removeLangGraphInterruptAction,
       }}
     >
       <CopilotMessages>{children}</CopilotMessages>